pop-up outline and design changes
Created an outline for the pop-up, in addition to connecting it to the settings button. The outline contains a header called settings, two checkboxes (not functional), and a close button (functional). All buttons have the same styling, which includes an enlargement, cursor pointer, and color change upon hover. Additionally, the toolbar colors and logo were adjusted to a less harsh shade of green. Colors are still subject to change along with fonts.